+++ packages/mochi-rest/src/modules/module.ts
@@ -3,4 +3,13 @@ import { Wretch } from "wretch/types";
 
 export class Module {
   constructor(public api: QueryStringAddon & Wretch<QueryStringAddon>) {}
+
+  token(t: string | null) {
+    if (t && this.api) {
+      this.api = this.api.auth(`Bearer ${t}`);
+      return;
+    }
+    this.api = this.api.auth("");
+    return;
+  }
 }